Remove (nearly) all new uses of `runtime.GOOS` by refactoring to use build tags instead.
