+++ cmd/secretgeneric_test.go
@@ -1,7 +1,6 @@
 package cmd
 
 import (
-	"runtime"
 	"testing"
 	"time"
 
@@ -10,12 +9,9 @@ import (
 
 func TestSecretFunc(t *testing.T) {
 	t.Parallel()
-	c := &Config{
-		GenericSecret: genericSecretCmdConfig{
-			Command: "date",
-		},
-	}
-	args := []string{"+%Y-%M-%DT%H:%M:%SZ"}
+
+	c, args := getSecretTestConfig()
+
 	var value interface{}
 	assert.NotPanics(t, func() {
 		value = c.secretFunc(args...)
@@ -27,26 +23,7 @@ func TestSecretFunc(t *testing.T) {
 func TestSecretJSONFunc(t *testing.T) {
 	t.Parallel()
 
-	var c *Config
-	var args []string
-
-	if runtime.GOOS != "windows" {
-		c = &Config{
-			GenericSecret: genericSecretCmdConfig{
-				Command: "date",
-			},
-		}
-		args = []string{`+{"date":"%Y-%M-%DT%H:%M:%SZ"}`}
-	} else {
-		// Windows doesn't (usually) have "date", but powershell is included with
-		// all versions of Windows v7 or newer.
-		c = &Config{
-			GenericSecret: genericSecretCmdConfig{
-				Command: "powershell.exe",
-			},
-		}
-		args = []string{"-Command", "Get-Date | ConvertTo-Json"}
-	}
+	c, args := getSecretJSONTestConfig()
 
 	var value interface{}
 	assert.NotPanics(t, func() {
